implement checking if the host is in Hosts table and fix tests
