+++ package/src/CacheManager.php
@@ -8,11 +8,13 @@ use Psr\Container\NotFoundExceptionInterface;
 use Sopamo\ClusterCache\Drivers\MemoryDriverInterface;
 use Sopamo\ClusterCache\Exceptions\CacheEntryValueIsOutOfMemoryException;
 use Sopamo\ClusterCache\Exceptions\NotFoundLocalCacheKeyException;
+use Sopamo\ClusterCache\Exceptions\PutCacheException;
 use Sopamo\ClusterCache\HostCommunication\Event;
 use Sopamo\ClusterCache\HostCommunication\HostCommunication;
 use Sopamo\ClusterCache\LockingMechanisms\DBLocker;
 use Sopamo\ClusterCache\LockingMechanisms\MemoryBlockLocker;
 use Sopamo\ClusterCache\Models\CacheEntry;
+use Sopamo\ClusterCache\Models\Host;
 
 class CacheManager
 {
@@ -36,8 +38,15 @@ class CacheManager
         $this->missBroadcastingForSpecialCacheKeys = [config('clustercache.prefix') . ':clustercache_hosts'];
     }
 
+    /**
+     * @throws PutCacheException
+     */
     public function put(string $key, mixed $value, int $ttl = 0): bool
     {
+        if(!Host::where('ip', HostHelpers::getHostIp())->exists()) {
+            throw new PutCacheException('Host is disconnected');
+        }
+
         if ($this->dbLocker->isLocked($key)) {
             return false;
         }
@@ -53,7 +62,7 @@ class CacheManager
                 ]
             );
 
-            //logger("Key: $key");
+            logger("Key: $key");
             //logger("missBroadcastingForSpecialCacheKeys: " . json_encode($this->missBroadcastingForSpecialCacheKeys));
             // It doesn't make sense to broadcast events like updating the internal cached data - for example "clustercache_hosts"
             if(!in_array($key, $this->missBroadcastingForSpecialCacheKeys)) {
@@ -70,13 +79,13 @@ class CacheManager
 
             $this->dbLocker->release($key);
 
-            //logger("putting successfully");
+            logger("putting successfully");
             return true;
         } catch (CacheEntryValueIsOutOfMemoryException $e) {
             logger($e->getMessage());
             $this->dbLocker->release($key);
 
-            return false;
+            throw new PutCacheException('Cache entry value is of out the memory');
         }
     }
