+++ package/src/Exceptions/PutCacheException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sopamo\ClusterCache\Exceptions;
+
+use Exception;
+
+class PutCacheException extends Exception
+{
+
+}
