+++ package/src/HostCommunication/HostCommunicationStatus.php
@@ -22,9 +22,9 @@ class HostCommunicationStatus
     }
 
     public static function leave():void {
+        Cache::store('clustercache')->put('clustercache_hosts', Host::where('ip', '!=', HostHelpers::getHostIp())->pluck('ip'));
         Host::where('ip', HostHelpers::getHostIp())->delete();
         DisconnectedHost::where('from', HostHelpers::getHostIp())->orWhere('to', HostHelpers::getHostIp())->delete();
-        Cache::store('clustercache')->put('clustercache_hosts', Host::pluck('ip'));
         app(HostCommunication::class)->triggerAll(Event::fromInt(Event::$allEvents['FETCH_HOSTS']));
     }
